+++ rules/variables.js
@@ -15,11 +15,11 @@ module.exports = {
     // disallow use of undefined when initializing variables
     'no-undef-init': 0,
     // disallow use of undeclared variables unless mentioned in a /*global */ block
-    'no-undef': 2,
+    'no-undef': 1,
     // disallow use of undefined variable
     'no-undefined': 0,
     // disallow declaration of variables that are not used in the code
-    'no-unused-vars': [2, { 'vars': 'local', 'args': 'after-used' }],
+    'no-unused-vars': [1, { 'vars': 'local', 'args': 'after-used' }],
     // disallow use of variables before they are defined
     'no-use-before-define': 2
   }